# config/swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "AlgorithemFinal",
      "version": "v1"
    },
    "paths": {
...
      "/api/Bells": {
        "get": {
          "tags": [
            "Bells"
          ],
          "parameters": [
            {
              "name": "PageSize",
              "in": "query",
              "description": "تعداد آیتم های موجود در صفحه",
              "schema": {
                "type": "string",
                "description": "تعداد آیتم های موجود در صفحه"
                
              }
            },
            {
              "name": "Page",
              "in": "query",
              "description": "شماره صفحه",
              "schema": {
                "type": "string",
                "description": "شماره صفحه"
                
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/BellPaginatedResult"
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "Bells"
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BellRequest"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BellRequest"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/BellRequest"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Bell"
                  }
                }
              }
            }
          }
        }
      },
      "/api/Bells/{id}": {
        "get": {
          "tags": [
            "Bells"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
                
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Bell"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "Bells"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
                
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BellRequest"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BellRequest"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/BellRequest"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        },
        "delete": {
          "tags": [
            "Bells"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
                
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
...
      "/api/Courses/{id}/TimeTables": {
        "get": {
          "tags": [
            "Courses"
          ],
          "summary": "برنامه زمانی درس هایی که این شناسه را دارند را برمیگرداند.",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "شناسه",
              "required": true,
              "schema": {
                "type": "string",
                "description": "شناسه"
                
              }
            },
            {
              "name": "PageSize",
              "in": "query",
              "description": "تعداد آیتم های موجود در صفحه",
              "schema": {
                "type": "string",
                "description": "تعداد آیتم های موجود در صفحه"
                
              }
            },
            {
              "name": "Page",
              "in": "query",
              "description": "شماره صفحه",
              "schema": {
                "type": "string",
                "description": "شماره صفحه"
                
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/TimeTablePaginatedResult"
                  }
                }
              }
            }
          }
        }
      },
...
      "/api/TimeTableBells": {
        "get": {
          "tags": [
            "TimeTableBells"
          ],
          "parameters": [
            {
              "name": "PageSize",
              "in": "query",
              "description": "تعداد آیتم های موجود در صفحه",
              "schema": {
                "type": "string",
                "description": "تعداد آیتم های موجود در صفحه"
                
              }
            },
            {
              "name": "Page",
              "in": "query",
              "description": "شماره صفحه",
              "schema": {
                "type": "string",
                "description": "شماره صفحه"
                
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/TimeTableBellPaginatedResult"
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "TimeTableBells"
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeTableBellRequest"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeTableBellRequest"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeTableBellRequest"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/TimeTableBell"
                  }
                }
              }
            }
          }
        }
      },
      "/api/TimeTableBells/{id}": {
        "get": {
          "tags": [
            "TimeTableBells"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
                
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/TimeTableBell"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "TimeTableBells"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
                
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
      "/api/TimeTables": {
        "get": {
          "tags": [
            "TimeTables"
          ],
          "summary": "",
          "parameters": [
            {
              "name": "StudentId",
              "in": "query",
              "description": "شناسه دانشجو",
              "schema": {
                "type": "string",
                "description": "شناسه دانشجو",
                "nullable": true
              }
            },
            {
              "name": "CourseId",
              "in": "query",
              "description": "شناسه درس",
              "schema": {
                "type": "string",
                "description": "شناسه درس"
                ,
                "nullable": true
              }
            },
            {
              "name": "MasterId",
              "in": "query",
              "description": "شناسه استاد",
              "schema": {
                "type": "string",
                "description": "شناسه استاد",
                "nullable": true
              }
            },
            {
              "name": "PageSize",
              "in": "query",
              "description": "تعداد آیتم های موجود در صفحه",
              "schema": {
                "type": "string",
                "description": "تعداد آیتم های موجود در صفحه"
                
              }
            },
            {
              "name": "Page",
              "in": "query",
              "description": "شماره صفحه",
              "schema": {
                "type": "string",
                "description": "شماره صفحه"
                
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/TimeTablePaginatedResult"
                  }
                }
              }
            }
          }
        }
      },
      "/api/TimeTables/{id}": {
        "get": {
          "tags": [
            "TimeTables"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
                
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/TimeTableDeatilsResponse"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
        "User": {
          "required": [
            "code"
          ],
          "type": "object",
          "properties": {
            "id": {
              "type": "string"
              
            },
            "lastName": {
              "type": "string",
              "nullable": true
            },
            "firstName": {
              "type": "string",
              "nullable": true
            },
            "code": {
              "type": "string",
              "description": "نام کاربری یا همان شماره دانشجویی"
            },
            "role": {
              "type": "string",
              "description": "master | admin | student",
              "nullable": true,
              "readOnly": true
            }
          },
          "additionalProperties": false
        },
        "Master": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string"
              
            },
            "userId": {
              "type": "string"
              
            },
            "user": {
              "$ref": "#/components/schemas/User"
            }
          },
          "additionalProperties": false
        },
        "Day": {
          "required": [
            "label"
          ],
          "type": "object",
          "properties": {
            "id": {
              "type": "string"
              
            },
            "label": {
              "type": "string"
            },
            "dayOfWeek": {
              "type": "string"
              
            }
          },
          "additionalProperties": false
        },
        "Bell": {
          "required": [
            "label"
          ],
          "type": "object",
          "properties": {
            "id": {
              "type": "string"
              
            },
            "label": {
              "type": "string"
            },
            "bellOfDay": {
              "type": "string"
              
            }
          },
          "additionalProperties": false
        },
        "TimeTableBell": {
          "required": [
            "bell",
            "day"
          ],
          "type": "object",
          "properties": {
            "id": {
              "type": "string"
              
            },
            "day": {
              "$ref": "#/components/schemas/Day"
            },
            "bell": {
              "$ref": "#/components/schemas/Bell"
            }
          },
          "additionalProperties": false
        },
        "Course": {
          "required": [
            "title"
          ],
          "type": "object",
          "properties": {
            "id": {
              "type": "string"
              
            },
            "title": {
              "type": "string"
            },
            "unitsCount": {
              "type": "string"
              
            }
          },
          "additionalProperties": false
        },
        "TimeTable": {
          "required": [
            "course"
          ],
          "type": "object",
          "properties": {
            "id": {
              "type": "string"
              
            },
            "master": {
              "$ref": "#/components/schemas/Master"
            },
            "timeTableBells": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/TimeTableBell"
              },
              "nullable": true
            },
            "course": {
              "$ref": "#/components/schemas/Course"
            }
          },
          "additionalProperties": false
        },
...
        "BellPaginatedResult": {
          "type": "object",
          "properties": {
            "list": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Bell"
              },
              "nullable": true
            },
            "count": {
              "type": "string"
              
            },
            "page": {
              "type": "string"
              
            },
            "totalPages": {
              "type": "string"
              
            }
          },
          "additionalProperties": false
        },
        "BellRequest": {
          "type": "object",
          "properties": {
            "label": {
              "type": "string",
              "nullable": true
            },
            "bellOfDay": {
              "type": "string"
              ,
              "nullable": true
            }
          },
          "additionalProperties": false
        },
...
        "TimeTablePaginatedResult": {
          "type": "object",
          "properties": {
            "list": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/TimeTable"
              },
              "nullable": true
            },
            "count": {
              "type": "string"
              
            },
            "page": {
              "type": "string"
              
            },
            "totalPages": {
              "type": "string"
              
            }
          },
          "additionalProperties": false
        },
...
        "TimeTableBellPaginatedResult": {
          "type": "object",
          "properties": {
            "list": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/TimeTableBell"
              },
              "nullable": true
            },
            "count": {
              "type": "string"
              
            },
            "page": {
              "type": "string"
              
            },
            "totalPages": {
              "type": "string"
            }
          },
          "additionalProperties": false
        },
        "TimeTableBellRequest": {
          "required": [
            "bellId",
            "dayId"
          ],
          "type": "object",
          "properties": {
            "dayId": {
              "type": "string",
              "description": "id rouz"            
            },
            "bellId": {
              "type": "string",
              "description": "id zang"
              
            }
          },
          "additionalProperties": false
        },
        "TimeTableDeatilsResponse": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string"
              
            },
            "master": {
              "$ref": "#/components/schemas/Master"
            },
            "timeTableBells": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/TimeTableBell"
              },
              "nullable": true
            },
            "course": {
              "$ref": "#/components/schemas/Course"
            }
          },
          "additionalProperties": false
        },
...
      },
      "securitySchemes": {
        "Token": {
          "type": "apiKey",
          "description": "JWT Authorization header using the Bearer scheme.",
          "name": "Authorization",
          "in": "header"
        }
      }
    },
    "security": [
      {
        "Token": [ ]
      }
    ]
  }